Extract CP855 to its own file
