+++ escpos4k/src/commonMain/kotlin/cz/multiplatform/escpos4k/core/encoding/charset/Charset.kt
@@ -88,141 +88,6 @@ public sealed class Charset(
   //  /** Page 32 [CP720: Arabic] */
   //  CP720(32),
 
-  /** Page 34 [CP855: Cyrillic] */
-  public data object CP855 : Charset(34, "IBM855") {
-    override val mapping: Map<Char, Int> =
-        mapOf(
-            '\u0452' to 0x80,
-            '\u0402' to 0x81,
-            '\u0453' to 0x82,
-            '\u0403' to 0x83,
-            '\u0451' to 0x84,
-            '\u0401' to 0x85,
-            '\u0454' to 0x86,
-            '\u0404' to 0x87,
-            '\u0455' to 0x88,
-            '\u0405' to 0x89,
-            '\u0456' to 0x8A,
-            '\u0406' to 0x8B,
-            '\u0457' to 0x8C,
-            '\u0407' to 0x8D,
-            '\u0458' to 0x8E,
-            '\u0408' to 0x8F,
-            '\u0459' to 0x90,
-            '\u0409' to 0x91,
-            '\u045A' to 0x92,
-            '\u040A' to 0x93,
-            '\u045B' to 0x94,
-            '\u040B' to 0x95,
-            '\u045C' to 0x96,
-            '\u040C' to 0x97,
-            '\u045E' to 0x98,
-            '\u040E' to 0x99,
-            '\u045F' to 0x9A,
-            '\u040F' to 0x9B,
-            '\u044E' to 0x9C,
-            '\u042E' to 0x9D,
-            '\u044A' to 0x9E,
-            '\u042A' to 0x9F,
-            '\u0430' to 0xA0,
-            '\u0410' to 0xA1,
-            '\u0431' to 0xA2,
-            '\u0411' to 0xA3,
-            '\u0446' to 0xA4,
-            '\u0426' to 0xA5,
-            '\u0434' to 0xA6,
-            '\u0414' to 0xA7,
-            '\u0435' to 0xA8,
-            '\u0415' to 0xA9,
-            '\u0444' to 0xAA,
-            '\u0424' to 0xAB,
-            '\u0433' to 0xAC,
-            '\u0413' to 0xAD,
-            '\u00AB' to 0xAE,
-            '\u00BB' to 0xAF,
-            '\u2591' to 0xB0,
-            '\u2592' to 0xB1,
-            '\u2593' to 0xB2,
-            '\u2502' to 0xB3,
-            '\u2524' to 0xB4,
-            '\u0445' to 0xB5,
-            '\u0425' to 0xB6,
-            '\u0438' to 0xB7,
-            '\u0418' to 0xB8,
-            '\u2563' to 0xB9,
-            '\u2551' to 0xBA,
-            '\u2557' to 0xBB,
-            '\u255D' to 0xBC,
-            '\u0439' to 0xBD,
-            '\u0419' to 0xBE,
-            '\u2510' to 0xBF,
-            '\u2514' to 0xC0,
-            '\u2534' to 0xC1,
-            '\u252C' to 0xC2,
-            '\u251C' to 0xC3,
-            '\u2500' to 0xC4,
-            '\u253C' to 0xC5,
-            '\u043A' to 0xC6,
-            '\u041A' to 0xC7,
-            '\u255A' to 0xC8,
-            '\u2554' to 0xC9,
-            '\u2569' to 0xCA,
-            '\u2566' to 0xCB,
-            '\u2560' to 0xCC,
-            '\u2550' to 0xCD,
-            '\u256C' to 0xCE,
-            '\u00A4' to 0xCF,
-            '\u043B' to 0xD0,
-            '\u041B' to 0xD1,
-            '\u043C' to 0xD2,
-            '\u041C' to 0xD3,
-            '\u043D' to 0xD4,
-            '\u041D' to 0xD5,
-            '\u043E' to 0xD6,
-            '\u041E' to 0xD7,
-            '\u043F' to 0xD8,
-            '\u2518' to 0xD9,
-            '\u250C' to 0xDA,
-            '\u2588' to 0xDB,
-            '\u2584' to 0xDC,
-            '\u041F' to 0xDD,
-            '\u044F' to 0xDE,
-            '\u2580' to 0xDF,
-            '\u042F' to 0xE0,
-            '\u0440' to 0xE1,
-            '\u0420' to 0xE2,
-            '\u0441' to 0xE3,
-            '\u0421' to 0xE4,
-            '\u0442' to 0xE5,
-            '\u0422' to 0xE6,
-            '\u0443' to 0xE7,
-            '\u0423' to 0xE8,
-            '\u0436' to 0xE9,
-            '\u0416' to 0xEA,
-            '\u0432' to 0xEB,
-            '\u0412' to 0xEC,
-            '\u044C' to 0xED,
-            '\u042C' to 0xEE,
-            '\u2116' to 0xEF,
-            '\u00AD' to 0xF0,
-            '\u044B' to 0xF1,
-            '\u042B' to 0xF2,
-            '\u0437' to 0xF3,
-            '\u0417' to 0xF4,
-            '\u0448' to 0xF5,
-            '\u0428' to 0xF6,
-            '\u044D' to 0xF7,
-            '\u042D' to 0xF8,
-            '\u0449' to 0xF9,
-            '\u0429' to 0xFA,
-            '\u0447' to 0xFB,
-            '\u0427' to 0xFC,
-            '\u00A7' to 0xFD,
-            '\u25A0' to 0xFE,
-            '\u00A0' to 0xFF,
-        )
-  }
-
   /** Page 35 [CP861: Icelandic] */
   public data object CP861 : Charset(35, "IBM861") {
     override val mapping: Map<Char, Int> =
